+++ utils/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/KafkaAvroBaseComponentTransformer.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.engine.avro
+
+object KafkaAvroBaseComponentTransformer {
+  final val SchemaVersionParamName = "Schema version"
+  final val TopicParamName = "Topic"
+  final val SinkKeyParamName = "Key"
+  final val SinkValueParamName = "Value"
+  final val SinkValidationModeParameterName = "Value validation mode"
+}