+++ utils/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/serialization/AbstractConfluentKafkaAvroDeserializer.scala
@@ -1,0 +1,47 @@
+package pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization
+
+import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException
+import io.confluent.kafka.serializers.{AbstractKafkaAvroDeserializer, AbstractKafkaSchemaSerDe}
+import org.apache.avro.io.DecoderFactory
+import org.apache.kafka.common.errors.SerializationException
+import pl.touk.nussknacker.engine.avro.RuntimeSchemaData
+import pl.touk.nussknacker.engine.avro.schema.{DatumReaderWriterMixin, RecordDeserializer}
+import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.ConfluentUtils
+
+import java.io.IOException
+import java.nio.ByteBuffer
+
+/**
+ * This class basically do the same as AbstractKafkaAvroDeserializer but use our createDatumReader implementation with time conversions
+ */
+abstract class AbstractConfluentKafkaAvroDeserializer extends AbstractKafkaAvroDeserializer with DatumReaderWriterMixin with RecordDeserializer {
+
+  override protected lazy val decoderFactory: DecoderFactory = DecoderFactory.get()
+
+  protected def deserialize(topic: String, isKey: java.lang.Boolean, payload: Array[Byte], readerSchema: Option[RuntimeSchemaData]): AnyRef = {
+    val buffer = ConfluentUtils.parsePayloadToByteBuffer(payload).valueOr(ex => throw ex)
+    read(buffer, readerSchema)
+  }
+
+  protected def read(buffer: ByteBuffer, expectedSchemaData: Option[RuntimeSchemaData]): AnyRef = {
+    var schemaId = -1
+
+    try {
+      schemaId = buffer.getInt
+      val parsedSchema = schemaRegistry.getSchemaById(schemaId)
+      val writerSchemaData = RuntimeSchemaData(ConfluentUtils.extractSchema(parsedSchema), Some(schemaId))
+      val readerSchemaData = expectedSchemaData.getOrElse(writerSchemaData)
+      // HERE we create our DatumReader
+      val reader = createDatumReader(writerSchemaData.schema, readerSchemaData.schema, useSchemaReflection, useSpecificAvroReader)
+      val bufferDataStart = 1 + AbstractKafkaSchemaSerDe.idSize
+      deserializeRecord(readerSchemaData, reader, buffer, bufferDataStart)
+    } catch {
+      case exc: RestClientException =>
+        throw new SerializationException(s"Error retrieving Avro schema for id : $schemaId", exc)
+      case exc@(_: RuntimeException | _: IOException) =>
+        // avro deserialization may throw IOException, AvroRuntimeException, NullPointerException, etc
+        throw new SerializationException(s"Error deserializing Avro message for id: $schemaId", exc)
+    }
+  }
+
+}